removed common core project
