--- src/NewLake.sln
@@ -1,63 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31912.275
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{83FE2D13-3924-4B12-A82A-BE09C1197478}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NewLake.Api", "services\NewLake.Api\NewLake.Api.csproj", "{CE3424A1-B6D7-4835-9BD2-CD9B60FCC8FF}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NewLake.Core", "services\NewLake.Core\NewLake.Core.csproj", "{30C79DAD-1923-4156-83B6-2E727E5C65E5}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NewLake.Queue.Listener", "services\NewLake.Queue.Listener\NewLake.Queue.Listener.csproj", "{488610FB-9A00-4983-A8A7-A086ABC2B600}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NewLake.GrpcServer", "services\NewLake.GrpcServer\NewLake.GrpcServer.csproj", "{42B2633E-8E72-4D21-B0F2-66B24A2A61FA}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5328E328-2CF5-4A9B-82A8-704A639CDF38}"
-	ProjectSection(SolutionItems) = preProject
-		..\README.md = ..\README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NewLake.Background.Api", "services\NewLake.Background.Api\NewLake.Background.Api.csproj", "{3E75150C-A0CC-4881-ABC2-7DCA95571638}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CE3424A1-B6D7-4835-9BD2-CD9B60FCC8FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CE3424A1-B6D7-4835-9BD2-CD9B60FCC8FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CE3424A1-B6D7-4835-9BD2-CD9B60FCC8FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CE3424A1-B6D7-4835-9BD2-CD9B60FCC8FF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{30C79DAD-1923-4156-83B6-2E727E5C65E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30C79DAD-1923-4156-83B6-2E727E5C65E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30C79DAD-1923-4156-83B6-2E727E5C65E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30C79DAD-1923-4156-83B6-2E727E5C65E5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{488610FB-9A00-4983-A8A7-A086ABC2B600}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{488610FB-9A00-4983-A8A7-A086ABC2B600}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{488610FB-9A00-4983-A8A7-A086ABC2B600}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{488610FB-9A00-4983-A8A7-A086ABC2B600}.Release|Any CPU.Build.0 = Release|Any CPU
-		{42B2633E-8E72-4D21-B0F2-66B24A2A61FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{42B2633E-8E72-4D21-B0F2-66B24A2A61FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{42B2633E-8E72-4D21-B0F2-66B24A2A61FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{42B2633E-8E72-4D21-B0F2-66B24A2A61FA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3E75150C-A0CC-4881-ABC2-7DCA95571638}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3E75150C-A0CC-4881-ABC2-7DCA95571638}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3E75150C-A0CC-4881-ABC2-7DCA95571638}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3E75150C-A0CC-4881-ABC2-7DCA95571638}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{CE3424A1-B6D7-4835-9BD2-CD9B60FCC8FF} = {83FE2D13-3924-4B12-A82A-BE09C1197478}
-		{30C79DAD-1923-4156-83B6-2E727E5C65E5} = {83FE2D13-3924-4B12-A82A-BE09C1197478}
-		{488610FB-9A00-4983-A8A7-A086ABC2B600} = {83FE2D13-3924-4B12-A82A-BE09C1197478}
-		{42B2633E-8E72-4D21-B0F2-66B24A2A61FA} = {83FE2D13-3924-4B12-A82A-BE09C1197478}
-		{3E75150C-A0CC-4881-ABC2-7DCA95571638} = {83FE2D13-3924-4B12-A82A-BE09C1197478}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1C7E4FCA-5BA7-4B8D-8FEA-65C580753CDE}
-	EndGlobalSection
-EndGlobal
