+++ src/services/NewLake.Api/Infrastructure/Services/Messaging/IMessageService.cs
@@ -1,0 +1,7 @@
+﻿namespace NewLake.Api.Infrastructure.Services
+{
+    public interface IMessageService<TMessage>
+    {
+        void Publish(TMessage message);
+    }
+}
