+++ src/services/NewLake.Api/Model/CacheItemBase.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace NewLake.Api.Model
+{
+    public abstract class CacheItemBase
+    {
+        public string Key { get; set; }
+        public string Value { get; set; }
+        public DateTime LastUpdated { get; set; }
+    }
+
+
+    public class CacheItem : CacheItemBase
+    {
+    }
+}
