+++ src/services/NewLake.Api/NewLake.Api.csproj
@@ -20,13 +20,19 @@
     <Folder Include="Infrastructure\Extensions\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\NewLake.Core\NewLake.Core.csproj">
-      <GlobalPropertiesToRemove></GlobalPropertiesToRemove>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
+    <PackageReference Include="RabbitMQ.Client" Version="6.2.4" />
+    <PackageReference Include="Grpc.AspNetCore" Version="2.44.0" />
+    <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="6.0.4" />
     <PackageReference Include="FluentValidation.AspNetCore" Version="11.1.2" />
     <PackageReference Include="Serilog.AspNetCore" Version="6.0.1" />
     <PackageReference Include="Serilog.Sinks.Seq" Version="5.1.1" />
+  </ItemGroup>
+   <ItemGroup>
+    <Protobuf Include="Proto\enums.proto" GrpcServices="None" ProtoRoot="Proto\">
+      <SubType>Component</SubType>
+    </Protobuf>
+    <Protobuf Include="Proto\bulkdata.proto" ProtoRoot="Proto\">
+      <SubType>Component</SubType>
+    </Protobuf>
   </ItemGroup>
 </Project>
