+++ src/services/NewLake.Api/Startup.cs
@@ -1,4 +1,6 @@
-﻿namespace NewLake.Api
+﻿
+
+namespace NewLake.Api
 {
     public class Startup
     {
