--- src/services/NewLake.Core/Infrastructure/QueueSettings.cs
@@ -1,11 +1,0 @@
-﻿namespace NewLake.Core.Infrastructure
-{
-    public class QueueSettings
-    {
-        public string QueueName { get; set; }
-        public string HostName { get; set; }
-        public string Exchange { get; set; }
-        public string Topic { get; set; }
-        public string RoutingKey { get; set; }
-    }
-}
