--- src/services/NewLake.Core/NewLake.Core.csproj
@@ -1,39 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Proto\bulkdata.proto" />
-    <None Remove="Proto\enums.proto" />
-    <None Remove="StackExchange.Redis.Extensions.AspNetCore" />
-    <None Remove="StackExchange.Redis.Extensions.System.Text.Json" />
-    <None Remove="RabbitMQ.Client" />
-    <None Remove="Azure.Security.KeyVault.Secrets" />
-    <None Remove="Azure.Identity" />
-    <None Remove="Microsoft.Extensions.Configuration.AzureAppConfiguration" />
-    <None Remove="Microsoft.Azure.AppConfiguration.AspNetCore" />
-    <None Remove="Services\" />
-    <None Remove="Services\Message\" />
-    <None Remove="Services\Caching\" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Grpc.AspNetCore" Version="2.44.0" />
-    <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="6.0.4" />
-    <PackageReference Include="RabbitMQ.Client" Version="6.2.4" />
-    <PackageReference Include="Azure.Identity" Version="1.6.0" />
-    <PackageReference Include="Microsoft.Azure.AppConfiguration.AspNetCore" Version="5.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Services\Caching\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Protobuf Include="Proto\enums.proto" GrpcServices="None" ProtoRoot="Proto\">
-      <SubType>Component</SubType>
-    </Protobuf>
-    <Protobuf Include="Proto\bulkdata.proto" ProtoRoot="Proto\">
-      <SubType>Component</SubType>
-    </Protobuf>
-  </ItemGroup>
-</Project>
