--- src/services/NewLake.Core/Services/Messaging/IMessageService.cs
@@ -1,7 +1,0 @@
-﻿namespace NewLake.Core.Services.Messaging
-{
-    public interface IMessageService<TMessage>
-    {
-        void Publish(TMessage message);
-    }
-}
