+++ src/services/NewLake.GrpcGenerator/.dockerignore
