+++ src/services/NewLake.GrpcGenerator/NewLake.GrpcGenerator.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk.Worker">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>    
+    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+    <DockerfileContext>..\..</DockerfileContext>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Google.Protobuf" Version="3.20.0" />
+    <PackageReference Include="Grpc.Net.ClientFactory" Version="2.44.0" />
+    <PackageReference Include="Grpc.Tools" Version="2.45.0">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />    
+  </ItemGroup>
+
+  <ItemGroup>
+    <Protobuf Include="Proto\enums.proto" GrpcServices="None" ProtoRoot="Proto\">
+      <Link>Proto\enums.proto</Link>
+    </Protobuf>
+    <Protobuf Include="Proto\bulkdata.proto" GrpcServices="Client" ProtoRoot="Proto\">
+      <Link>Proto\bulkdata.proto</Link>
+    </Protobuf>
+  </ItemGroup>
+</Project>
