+++ src/services/NewLake.GrpcGenerator/ServiceSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace NewLake.GrpcGenerator
+{
+    public class ServiceSettings
+    {
+        public int DelayInterval { get; set; }
+        public int MessageId { get; set; }
+        public string? ServerUrl { get; set; }
+    }
+}
