+++ src/services/NewLake.GrpcGenerator/Services/IBulkPacketGenerator.cs
@@ -1,0 +1,9 @@
+﻿using NewLake.GrpcGenerator.Services;
+
+namespace NewLake.GrpcGenerator.Services
+{
+    public interface IBulkPacketGenerator
+    {
+        MessagePacket BuildMessagePacket(int packetId);
+    }
+}
