+++ src/services/NewLake.GrpcGenerator/appsettings.Development.json
@@ -1,0 +1,13 @@
+{
+  "ServiceSettings": {
+    "DelayInterval": 10000,
+    "MessageId": 1,    
+    "ServerUrl": "http://localhost:5002"
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  }
+}
