--- src/services/NewLake.GrpcServer/NewLake.GrpcServer.csproj
@@ -1,31 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Worker">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <UserSecretsId>dotnet-NewLake.GrpcMessage.Sender-9063CA42-366E-4509-B930-D58372F13484</UserSecretsId>
-    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-    <DockerfileContext>..\..</DockerfileContext>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Google.Protobuf" Version="3.20.0" />
-    <PackageReference Include="Grpc.Net.ClientFactory" Version="2.44.0" />
-    <PackageReference Include="Grpc.Tools" Version="2.45.0">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.15.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Protobuf Include="..\NewLake.Core\Proto\bulkdata.proto" GrpcServices="Client">
-      <Link>Protos\BulkData.proto</Link>
-    </Protobuf>
-    <Protobuf Include="..\NewLake.Core\Proto\enums.proto" GrpcServices="None">
-      <Link>Protos\enums.proto</Link>
-    </Protobuf>
-  </ItemGroup>
-</Project>
