--- src/services/NewLake.GrpcServer/ServiceSettings.cs
@@ -1,9 +1,0 @@
-﻿namespace NewLake.GrpcServer
-{
-    public class ServiceSettings
-    {
-        public int DelayInterval { get; set; }
-        public int MessageId { get; set; }
-        public string? ServerUrl { get; set; }
-    }
-}
