--- src/services/NewLake.GrpcServer/Services/IBulkPacketGenerator.cs
@@ -1,9 +1,0 @@
-﻿using NewLake.Core.GrpcProto.Services;
-
-namespace NewLake.GrpcServer.Sender.Services
-{
-    public interface IBulkPacketGenerator
-    {
-        MessagePacket BuildMessagePacket(int packetId);
-    }
-}
