--- src/services/NewLake.GrpcServer/appsettings.Development.json
@@ -1,13 +1,0 @@
-{
-  "ServiceSettings": {
-    "DelayInterval": 10000,
-    "MessageId": 1,
-    "ServerUrl": "http://192.168.49.2:31502"
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  }
-}
