+++ src/services/NewLake.Queue.Listener/Infrastructure/QueueSettings.cs
@@ -1,0 +1,11 @@
+﻿namespace NewLake.Queue.Listener.Infrastructure.Settings
+{
+    public class QueueSettings
+    {
+        public string QueueName { get; set; }
+        public string HostName { get; set; }
+        public string Exchange { get; set; }
+        public string Topic { get; set; }
+        public string RoutingKey { get; set; }
+    }
+}
